pyproject.toml
[project]
name = "mrc-discord-bot"
version = "0.1.0"
description = "Discord bot for MRC stat tracking"
readme = "README.md"
requires-python = ">=3.12,<3.13"
dependencies = [
  "discord.py>=2.4.0,<3",
  "numpy>=1.26.0",
  "opencv-python-headless>=4.12.0.88",
  "paddleocr==3.3.2",
  "paddlepaddle==3.2.2",
  "paddlex[ocr]==3.3.10",
  "pillow>=10.0.0",
  "python-dotenv>=1.0.0",
]

[dependency-groups]
dev = []
